+++ react/index.tsx
@@ -1,12 +1,9 @@
 import React, { Component } from 'react'
+import 'vtex.country-codes/locales'
 
 import AppRouter from './components/AppRouter'
 import ClientSide from './components/ClientSide'
 import { logMyAccountURL, logGeneralErrors } from './utils/splunk'
-
-import 'vtex.country-codes/locales'
-
-import styles from './styles.css'
 
 class MyAccount extends Component {
   public constructor(props: any) {
