+++ react/typings/vtex.render-runtime.d.ts
@@ -1,4 +1,4 @@
-declare module 'render' {
+declare module 'vtex.render-runtime' {
   export const ExtensionPoint
   export const withRuntimeContext
 }
